US #11 Validation Errors

+++ grails-app/controllers/net/kaleidos/plugins/admin/GrailsAdminPluginApiController.groovy
@@ -1,6 +1,7 @@
 package net.kaleidos.plugins.admin
 
 import grails.converters.JSON
+import grails.validation.ValidationException
 
 class GrailsAdminPluginApiController {
     def objectDefinitionSource
@@ -11,12 +12,12 @@ class GrailsAdminPluginApiController {
         render adminConfigHolder.domainClasses as JSON
     }
 
-    def getAdminAction(String domain, Long id) {
-        def config = _resolve(domain)
+    def getAdminAction(String slug, Long id) {
+        def config = adminConfigHolder.getDomainConfigBySlug(slug)
         if (!config) {
             response.status = 404
             render(["error":"Domain no configured"] as JSON)
-            return;
+            return
         }
 
         def result
@@ -25,7 +26,7 @@ class GrailsAdminPluginApiController {
             if (!result) {
                 response.status = 404
                 render(["error":"Entity not found"] as JSON)
-                return;
+                return
             }
         } else {
             result = grailsAdminPluginGenericService.listDomains(config.domainClass.clazz)
@@ -33,32 +34,62 @@ class GrailsAdminPluginApiController {
         render result as JSON
     }
 
-    def putAdminAction(String domain) {
-        def config = _resolve(domain)
-        def result = grailsAdminPluginGenericService.saveDomain(config.domainClass.clazz, request.JSON)
+    def putAdminAction(String slug) {
+        def config = adminConfigHolder.getDomainConfigBySlug(slug)
+        if (!config) {
+            response.status = 404
+            render(["error":"Domain no configured"] as JSON)
+            return
+        }
+
+        def result = [:]
+        try {
+            result = grailsAdminPluginGenericService.saveDomain(config.domainClass.clazz, request.JSON)
+        } catch (ValidationException e) {
+            response.status = 500
+            result = e.getErrors()
+        }
         render result as JSON
     }
 
-    def postAdminAction(String domain, Long id) {
-        def config = _resolve(domain)
-        def result = grailsAdminPluginGenericService.updateDomain(config.domainClass.clazz, id, request.JSON)
+    def postAdminAction(String slug, Long id) {
+        def config = adminConfigHolder.getDomainConfigBySlug(slug)
+        if (!config) {
+            response.status = 404
+            render(["error":"Domain no configured"] as JSON)
+            return
+        }
+
+        def result = [:]
+        try {
+            result = grailsAdminPluginGenericService.updateDomain(config.domainClass.clazz, id, request.JSON)
+        } catch (ValidationException e) {
+            response.status = 500
+            result = e.getErrors()
+        } catch (RuntimeException e) {
+            response.status = 500
+            result = [error: e.message]
+        }
         render result as JSON
     }
 
-    def deleteAdminAction(String domain, Long id) {
-        def config = _resolve(domain)
-        grailsAdminPluginGenericService.deleteDomain(config.domainClass.clazz, id)
+    def deleteAdminAction(String slug, Long id) {
+        def config = adminConfigHolder.getDomainConfigBySlug(slug)
+        if (!config) {
+            response.status = 404
+            render(["error":"Domain no configured"] as JSON)
+            return
+        }
+
+        try {
+            grailsAdminPluginGenericService.deleteDomain(config.domainClass.clazz, id)
+        } catch (RuntimeException e) {
+            response.status = 500
+            def result = [error: e.message]
+            render result as JSON
+            return
+        }
         response.status = 204
         render ""
-    }
-
-    private _resolve(String adminController) {
-        def domainClasses  = adminConfigHolder.domainClasses
-        def className = domainClasses.find {
-            def className = it.tokenize(".")[-1]
-            def camelClassName = className[0].toLowerCase() + className.substring(1)
-            return camelClassName == adminController
-        }
-        return adminConfigHolder.domains[className]
     }
 }
